requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
tweepy>=4.14.0
tiktoken==0.7.0
chardet>=5.0.0
asyncio>=3.4.3
aiofiles>=23.2.1
gitpython>=3.1.40
tqdm>=4.66.1
transformers>=4.39.0
tokenizers>=0.19.1
anthropic>=0.37.1,<0.38.0
psutil==5.9.6
aider-chat>=0.14.0
cairosvg>=2.7.1
packaging>=21.3  # For version comparison
